pyproject.toml
[tool.poetry]
name = "project-drf"
version = "0.1.0"
description = ""
authors = ["Pavel Akulich <pavelakulich1999@gmail.com>"]
readme = "README.md"
packages = [{include = "project_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
django-stubs = "^4.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
